pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Mikhail Egorov <pe.egorov@gmail.com>"]
readme = "README.md"
packages = [
  { include = "page_analyzer" },
]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.2"
python-dotenv = "^1.0.1"
flake8 = "^7.0.0"
gunicorn = "^21.2.0"
pytest = "^8.0.2"
coverage = "^7.4.3"
pytest-cov = "^4.1.0"
psycopg2-binary = "^2.9.9"
validators = "^0.22.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
